a2a-backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "a2a-backend"
version = "0.1.0"
description = "Bridge between assistant-transport protocol and A2A servers"
authors = [
    { name = "assistant-ui Team", email = "team@assistant-ui.com" }
]
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "assistant-stream>=0.0.28",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "a2a-sdk>=0.3.0",
    "httpx>=0.28.0",
]

[project.scripts]
a2a-backend = "main:main"

[tool.hatch.build.targets.wheel]
packages = ["."]
